reorg and search by description and tags
