+++ frontend/src/scenes/dashboard/dashboards/DashboardsTable.tsx
@@ -1,50 +1,34 @@
 import { useActions, useValues } from 'kea'
 import { dashboardsModel } from '~/models/dashboardsModel'
-import { Card, Tabs } from 'antd'
-import { dashboardsLogic, DashboardsTab } from 'scenes/dashboard/dashboardsLogic'
-import { Link } from 'lib/components/Link'
-import { AppstoreAddOutlined, PushpinFilled, PushpinOutlined, ShareAltOutlined } from '@ant-design/icons'
-import { NewDashboardModal } from 'scenes/dashboard/NewDashboardModal'
-import { PageHeader } from 'lib/components/PageHeader'
-import { AvailableFeature, DashboardMode, DashboardType } from '~/types'
-import { ObjectTags } from 'lib/components/ObjectTags/ObjectTags'
+import { dashboardsLogic, DashboardsTab } from 'scenes/dashboard/dashboards/dashboardsLogic'
 import { userLogic } from 'scenes/userLogic'
-import { DashboardEventSource } from 'lib/utils/eventUsageLogic'
-import { urls } from 'scenes/urls'
-import { SceneExport } from 'scenes/sceneTypes'
+import { teamLogic } from 'scenes/teamLogic'
+import { duplicateDashboardLogic } from 'scenes/dashboard/duplicateDashboardLogic'
+import { deleteDashboardLogic } from 'scenes/dashboard/deleteDashboardLogic'
 import { LemonTable, LemonTableColumn, LemonTableColumns } from 'lib/components/LemonTable'
-import { createdAtColumn, createdByColumn } from 'lib/components/LemonTable/columnUtils'
+import { AvailableFeature, DashboardMode, DashboardType } from '~/types'
 import { LemonButton } from 'lib/components/LemonButton'
-import { More } from 'lib/components/LemonButton/More'
-import { dashboardLogic } from './dashboardLogic'
-import { LemonRow } from 'lib/components/LemonRow'
-import { LemonDivider } from 'lib/components/LemonDivider'
+import { DashboardEventSource } from 'lib/utils/eventUsageLogic'
+import { PushpinFilled, PushpinOutlined, ShareAltOutlined } from '@ant-design/icons'
+import { DashboardPrivilegeLevel } from 'lib/constants'
+import { Link } from 'lib/components/Link'
+import { urls } from 'scenes/urls'
 import { Tooltip } from 'lib/components/Tooltip'
 import { IconCottage, IconLock } from 'lib/components/icons'
-import { teamLogic } from 'scenes/teamLogic'
-import { newDashboardLogic } from 'scenes/dashboard/newDashboardLogic'
-import { DashboardPrivilegeLevel } from 'lib/constants'
-import { inAppPromptLogic } from 'lib/logic/inAppPrompt/inAppPromptLogic'
-import { LemonInput } from '@posthog/lemon-ui'
-import { deleteDashboardLogic } from 'scenes/dashboard/deleteDashboardLogic'
-import { DeleteDashboardModal } from 'scenes/dashboard/DeleteDashboardModal'
-import { DuplicateDashboardModal } from 'scenes/dashboard/DuplicateDashboardModal'
-import { duplicateDashboardLogic } from 'scenes/dashboard/duplicateDashboardLogic'
+import { ObjectTags } from 'lib/components/ObjectTags/ObjectTags'
+import { createdAtColumn, createdByColumn } from 'lib/components/LemonTable/columnUtils'
+import { More } from 'lib/components/LemonButton/More'
+import { dashboardLogic } from 'scenes/dashboard/dashboardLogic'
+import { LemonDivider } from 'lib/components/LemonDivider'
+import { LemonRow } from 'lib/components/LemonRow'
 
-export const scene: SceneExport = {
-    component: Dashboards,
-    logic: dashboardsLogic,
-}
-
-export function Dashboards(): JSX.Element {
+export const DashboardsTable = (): JSX.Element => {
     const { dashboardsLoading } = useValues(dashboardsModel)
     const { unpinDashboard, pinDashboard } = useActions(dashboardsModel)
-    const { setSearchTerm, setCurrentTab } = useActions(dashboardsLogic)
+    const { setCurrentTab } = useActions(dashboardsLogic)
     const { dashboards, searchTerm, currentTab } = useValues(dashboardsLogic)
-    const { showNewDashboardModal, addDashboard } = useActions(newDashboardLogic)
     const { hasAvailableFeature } = useValues(userLogic)
     const { currentTeam } = useValues(teamLogic)
-    const { closePrompts } = useActions(inAppPromptLogic)
     const { showDuplicateDashboardModal } = useActions(duplicateDashboardLogic)
     const { showDeleteDashboardModal } = useActions(deleteDashboardLogic)
 
@@ -83,7 +67,13 @@ export function Dashboards(): JSX.Element {
                             </Link>
                             {!canEditDashboard && (
                                 <Tooltip title="You don't have edit permissions for this dashboard.">
-                                    <IconLock style={{ marginLeft: 6, verticalAlign: '-0.125em', display: 'inline' }} />
+                                    <IconLock
+                                        style={{
+                                            marginLeft: 6,
+                                            verticalAlign: '-0.125em',
+                                            display: 'inline',
+                                        }}
+                                    />
                                 </Tooltip>
                             )}
                             {is_shared && (
@@ -196,118 +186,36 @@ export function Dashboards(): JSX.Element {
     ]
 
     return (
-        <div>
-            <NewDashboardModal />
-            <DuplicateDashboardModal />
-            <DeleteDashboardModal />
-            <PageHeader
-                title="Dashboards"
-                buttons={
-                    <LemonButton
-                        data-attr={'new-dashboard'}
-                        onClick={() => {
-                            closePrompts()
-                            showNewDashboardModal()
-                        }}
-                        type="primary"
-                    >
-                        New dashboard
-                    </LemonButton>
-                }
-            />
-            <Tabs
-                activeKey={currentTab}
-                style={{ borderColor: '#D9D9D9' }}
-                onChange={(tab) => setCurrentTab(tab as DashboardsTab)}
-            >
-                <Tabs.TabPane tab="All Dashboards" key={DashboardsTab.All} />
-                <Tabs.TabPane tab="Your Dashboards" key={DashboardsTab.Yours} />
-                <Tabs.TabPane tab="Pinned" key={DashboardsTab.Pinned} />
-                <Tabs.TabPane tab="Shared" key={DashboardsTab.Shared} />
-            </Tabs>
-            <div className="flex">
-                <LemonInput
-                    type="search"
-                    placeholder="Search for dashboards"
-                    onChange={setSearchTerm}
-                    value={searchTerm}
-                />
-                <div />
-            </div>
-            <LemonDivider className="my-4" />
-            {dashboardsLoading || dashboards.length > 0 || searchTerm || currentTab !== DashboardsTab.All ? (
-                <LemonTable
-                    data-attr="dashboards-table"
-                    pagination={{ pageSize: 100 }}
-                    dataSource={dashboards}
-                    rowKey="id"
-                    columns={columns}
-                    loading={dashboardsLoading}
-                    defaultSorting={{ columnKey: 'name', order: 1 }}
-                    emptyState={
-                        searchTerm ? (
-                            `No ${
-                                currentTab === DashboardsTab.Pinned
-                                    ? 'pinned '
-                                    : currentTab === DashboardsTab.Shared
-                                    ? 'shared '
-                                    : ''
-                            }dashboards matching "${searchTerm}"!`
-                        ) : currentTab === DashboardsTab.Pinned ? (
-                            <>
-                                No dashboards have been pinned for quick access yet.{' '}
-                                <Link onClick={() => setCurrentTab(DashboardsTab.All)}>
-                                    Go to All Dashboards to pin one.
-                                </Link>
-                            </>
-                        ) : currentTab === DashboardsTab.Shared ? (
-                            <>
-                                No dashboards have been shared yet.{' '}
-                                <Link onClick={() => setCurrentTab(DashboardsTab.All)}>
-                                    Go to All Dashboards to share one.
-                                </Link>
-                            </>
-                        ) : undefined
-                    }
-                    nouns={['dashboard', 'dashboards']}
-                />
-            ) : (
-                <div className="mt-4">
-                    <p>Create your first dashboard:</p>
-                    <div className="flex justify-center items-center gap-4">
-                        <Card
-                            title="Empty"
-                            size="small"
-                            style={{ width: 200, cursor: 'pointer' }}
-                            onClick={() =>
-                                addDashboard({
-                                    name: 'New Dashboard',
-                                    useTemplate: '',
-                                })
-                            }
-                        >
-                            <div style={{ textAlign: 'center', fontSize: 40 }}>
-                                <AppstoreAddOutlined />
-                            </div>
-                        </Card>
-                        <Card
-                            title="App Default"
-                            size="small"
-                            style={{ width: 200, cursor: 'pointer' }}
-                            onClick={() =>
-                                addDashboard({
-                                    name: 'Web App Dashboard',
-                                    useTemplate: 'DEFAULT_APP',
-                                })
-                            }
-                        >
-                            <div style={{ textAlign: 'center', fontSize: 40 }}>
-                                <AppstoreAddOutlined />
-                            </div>
-                        </Card>
-                    </div>
-                </div>
-            )}
-        </div>
+        <LemonTable
+            data-attr="dashboards-table"
+            pagination={{ pageSize: 100 }}
+            dataSource={dashboards as DashboardType[]}
+            rowKey="id"
+            columns={columns}
+            loading={dashboardsLoading}
+            defaultSorting={{ columnKey: 'name', order: 1 }}
+            emptyState={
+                searchTerm ? (
+                    `No ${
+                        currentTab === DashboardsTab.Pinned
+                            ? 'pinned '
+                            : currentTab === DashboardsTab.Shared
+                            ? 'shared '
+                            : ''
+                    }dashboards matching "${searchTerm}"!`
+                ) : currentTab === DashboardsTab.Pinned ? (
+                    <>
+                        No dashboards have been pinned for quick access yet.{' '}
+                        <Link onClick={() => setCurrentTab(DashboardsTab.All)}>Go to All Dashboards to pin one.</Link>
+                    </>
+                ) : currentTab === DashboardsTab.Shared ? (
+                    <>
+                        No dashboards have been shared yet.{' '}
+                        <Link onClick={() => setCurrentTab(DashboardsTab.All)}>Go to All Dashboards to share one.</Link>
+                    </>
+                ) : undefined
+            }
+            nouns={['dashboard', 'dashboards']}
+        />
     )
 }
